chore: extract createCodexStateDb helper and update eslint-config
- Extract createCodexStateDb to lib/testing/codex-db.ts (shared by
  session.test.ts and cleanup.test.ts) to avoid schema duplication
  and fix positional INSERT in cleanup.test.ts
- Remove local n/no-unsupported-features/node-builtins override;
  the rule is now handled by @nozomiishii/eslint-config 1.9.0
- Fix new unicorn/comment-content and unicorn/prefer-await lint
  issues from eslint-config upgrade
