+++ package.json
@@ -41,7 +41,7 @@
   },
   "devDependencies": {
     "@nozomiishii/commitlint-config": "1.7.0",
-    "@nozomiishii/eslint-config": "1.7.0",
+    "@nozomiishii/eslint-config": "1.9.0",
     "@nozomiishii/lefthook-config": "1.7.0",
     "@nozomiishii/prettier-config": "1.7.0",
     "@nozomiishii/tsconfig": "1.7.0",
